requirements.txt
numpy>=1.21.0
pandas>=1.3.0
tensorflow>=2.10.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
python-dotenv>=0.19.0
joblib>=1.1.0
requests>=2.26.0
streamlit>=1.22.0
plotly>=5.13.0